requirements.txt
defusedxml==0.7.1
et-xmlfile==1.1.0
fonttools==4.54.1
fpdf2==2.7.9
openpyxl==3.1.5
pillow==10.4.0
